+++ pipelex/cogt/img_gen/img_gen_job_components.py
@@ -15,10 +15,14 @@ class AspectRatio(StrEnum):
     LANDSCAPE_3_2 = "landscape_3_2"
     LANDSCAPE_16_9 = "landscape_16_9"
     LANDSCAPE_21_9 = "landscape_21_9"
+    LANDSCAPE_4_1 = "landscape_4_1"
+    LANDSCAPE_8_1 = "landscape_8_1"
     PORTRAIT_3_4 = "portrait_3_4"
     PORTRAIT_2_3 = "portrait_2_3"
     PORTRAIT_9_16 = "portrait_9_16"
     PORTRAIT_9_21 = "portrait_9_21"
+    PORTRAIT_1_4 = "portrait_1_4"
+    PORTRAIT_1_8 = "portrait_1_8"
 
 
 class Quality(StrEnum):